Implemented the reading from file
struct FileCIDSource now contains the file, the filename and the scanner used to read the file.
The idea is that the GetNewCid() function will use the current state of the scanner to read the next CID along with it's contents.
I have doubts about using the parse function here to convert the string read from the file to a CID